pyproject.toml
[project]
name = "ghagen"
version = "0.2.1"
description = "Generate GitHub Actions workflow YAML from Python code"
readme = "README.md"
authors = [
    { name = "Nathan Jordan", email = "nathan.m.jordan@gmail.com" },
]
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["github-actions", "ci", "cd", "yaml", "workflow", "code-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Code Generators",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.6",
    "ruamel.yaml>=0.18",
    "tomli_w>=1.0",
    "typer>=0.12",
]

[project.urls]
Repository = "https://github.com/nathanjordan/ghagen"

[project.scripts]
ghagen = "ghagen.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages/python/src/ghagen"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-snapshot",
    "ruff",
    "pyright",
    "httpx",
    "datamodel-code-generator>=0.25",
    "jsonschema",
    "mkdocs-material",
    "mkdocstrings[python]",
]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["packages/python/src"]
extend-exclude = [
    "packages/python/src/ghagen/schema/snapshot/_generated_models.py",
    "packages/python/src/ghagen/schema/snapshot/_generated_action_models.py",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"
include = ["packages/python/src"]
exclude = [
    "packages/python/src/ghagen/schema/snapshot/_generated_models.py",
    "packages/python/src/ghagen/schema/snapshot/_generated_action_models.py",
]

[tool.pytest.ini_options]
testpaths = ["packages/python/tests"]
